requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
pymupdf
tqdm
requests
python-dotenv